Update: 25.02
Hello there guys.
In this update we resolved a lot of serious problems that
creates really, really bad situations in some parts of the project. We
have still unfinished problems so please check the M1.txt file and see
what we have done and what we have to do till the dead line (15.03.2015)